Initial commit
This implementation of robot uses its own framework of movement options and path algorithmns.
This will be replaced with the kinematic modelling of the robot and closed loop control
for path algorithm due to the original being unfeasible.

Also option to add and remove sensors from the robot is added along with other helper functions